Tidy up spoonacular api related scripts
This version does not function becuase if I fix it, git does not know
that it's the same file being moved to some place else. I need to keep
the content same to make git know that I'm just moving the file, not
creating a fresh new file. This helps with keeping the git history.